moved generic Filter class to correct folder
